ref: contained container commands in docker.go
